% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_list_hazard_inventory.R
\name{list_hazard_inventory}
\alias{list_hazard_inventory}
\title{List hazard inventory from loaded hazards}
\usage{
list_hazard_inventory(hazards)
}
\arguments{
\item{hazards}{list of SpatRaster as returned by load_hazards()}
}
\value{
data.frame with columns: key, hazard_type, scenario
}
\description{
List hazard inventory from loaded hazards
}
