% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_company_analysis.R
\name{mod_company_analysis_ui}
\alias{mod_company_analysis_ui}
\title{company_analysis UI Function}
\usage{
mod_company_analysis_ui(id)
}
\arguments{
\item{id}{Internal parameter for shiny}
}
\description{
Module to display company-level analysis with visualizations
}
