% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_control.R
\name{mod_control_ui}
\alias{mod_control_ui}
\title{control UI Function}
\usage{
mod_control_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}}
}
\description{
Sidebar control module for company upload, hazard selection, and analysis controls
}
