% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hazard_maps.R
\name{mod_hazard_maps_server}
\alias{mod_hazard_maps_server}
\title{hazard_maps Server Functions}
\usage{
mod_hazard_maps_server(
  id,
  results_reactive,
  events_reactive,
  hazards_reactive,
  base_dir_reactive
)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{results_reactive}{reactive containing analysis results}

\item{events_reactive}{reactive containing configured events}

\item{hazards_reactive}{reactive containing hazard data}

\item{base_dir_reactive}{reactive containing base directory path}
}
\description{
hazard_maps Server Functions
}
