% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hazard_maps.R
\name{mod_hazard_maps_ui}
\alias{mod_hazard_maps_ui}
\title{hazard_maps UI Function}
\usage{
mod_hazard_maps_ui(id)
}
\arguments{
\item{id}{Internal parameter for shiny}
}
\description{
Module to display interactive hazard maps with asset overlays
}
