% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hazards_events.R
\name{mod_hazards_events_ui}
\alias{mod_hazards_events_ui}
\title{hazards_events UI Function}
\usage{
mod_hazards_events_ui(id, title = "Hazard events")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}}

\item{title}{Character title displayed above the controls}
}
\description{
Shiny module to build a list of hazard events to apply. Allows adding
multiple events with type, hazard name, event year, and chronic toggle.
}
