% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profit_pathways.R
\name{mod_profit_pathways_ui}
\alias{mod_profit_pathways_ui}
\title{profit_pathways UI Function}
\usage{
mod_profit_pathways_ui(id)
}
\arguments{
\item{id}{Internal parameter for shiny}
}
\description{
Module to display asset profit pathways with interactive highlighting
}
