% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_assets.R
\name{mod_results_assets_ui}
\alias{mod_results_assets_ui}
\title{results_assets UI Function}
\usage{
mod_results_assets_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}}
}
\description{
Module to display asset-level results table with exposures
}
