% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_companies.R
\name{mod_results_companies_server}
\alias{mod_results_companies_server}
\title{results_companies Server Functions}
\usage{
mod_results_companies_server(id, results_reactive)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{results_reactive}{reactive containing analysis results}
}
\description{
results_companies Server Functions
}
