% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_companies.R
\name{mod_results_companies_ui}
\alias{mod_results_companies_ui}
\title{results_companies UI Function}
\usage{
mod_results_companies_ui(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}}
}
\description{
Module to display company-level results table with financials
}
