% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_server}
\alias{mod_settings_server}
\title{settings Server Functions}
\usage{
mod_settings_server(id, base_dir_reactive, hazard_configs_reactive)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{base_dir_reactive}{reactive containing base directory path}

\item{hazard_configs_reactive}{reactive list from load_hazard_configs()}
}
\value{
list with reload_trigger reactive counter
}
\description{
settings Server Functions
}
