% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_ui}
\alias{mod_settings_ui}
\title{settings UI Function}
\usage{
mod_settings_ui(id)
}
\arguments{
\item{id}{Internal parameter for shiny}
}
\description{
Settings module for hazard configuration overrides
}
