% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_status.R
\name{mod_status_server}
\alias{mod_status_server}
\title{status Server Functions}
\usage{
mod_status_server(
  id,
  status_reactive,
  events_reactive,
  delete_event_callback = NULL,
  run_repro_code_reactive = NULL,
  spatial_data_reactive = NULL
)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{status_reactive}{reactive containing current status message}

\item{events_reactive}{reactive containing configured events}

\item{delete_event_callback}{function to delete an event by event_id}

\item{run_repro_code_reactive}{optional reactive containing the generated run reproduction code}

\item{spatial_data_reactive}{optional reactive containing loaded spatial separation layers}
}
\description{
status Server Functions
}
