% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_status.R
\name{mod_status_ui}
\alias{mod_status_ui}
\title{status UI Function}
\usage{
mod_status_ui(id)
}
\arguments{
\item{id}{Internal parameter for shiny}
}
\description{
Module to display analysis status and configured events
}
