% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_company_analysis.R
\name{mod_company_analysis_server}
\alias{mod_company_analysis_server}
\title{company_analysis Server Functions}
\usage{
mod_company_analysis_server(id, results_reactive)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{results_reactive}{reactive containing analysis results}
}
\description{
company_analysis Server Functions
}
