% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_control.R
\name{mod_control_server}
\alias{mod_control_server}
\title{control Server Functions}
\usage{
mod_control_server(id, base_dir_reactive)
}
\arguments{
\item{base_dir_reactive}{reactive containing base directory path}
}
\value{
list with reactive values for company_file, events, run_trigger, results_ready, aggregation_factor, and get_hazards_at_factor
}
\description{
control Server Functions
}
