% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profit_pathways.R
\name{mod_profit_pathways_server}
\alias{mod_profit_pathways_server}
\title{profit_pathways Server Functions}
\usage{
mod_profit_pathways_server(id, results_reactive)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{results_reactive}{reactive containing analysis results}
}
\description{
profit_pathways Server Functions
}
