% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_status.R
\name{mod_status_server}
\alias{mod_status_server}
\title{status Server Functions}
\usage{
mod_status_server(
  id,
  status_reactive,
  events_reactive,
  results_reactive = NULL,
  delete_event_callback = NULL
)
}
\arguments{
\item{id}{Internal parameter for shiny}

\item{status_reactive}{reactive containing current status message}

\item{events_reactive}{reactive containing configured events}

\item{results_reactive}{reactive containing analysis results}

\item{delete_event_callback}{Function to call when deleting an event (optional)}
}
\description{
status Server Functions
}
