% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_status.R
\name{mod_status_server}
\alias{mod_status_server}
\title{status Server Functions}
\usage{
mod_status_server(id, status_reactive, events_reactive)
}
\arguments{
\item{status_reactive}{reactive containing current status message}

\item{events_reactive}{reactive containing configured events}
}
\description{
status Server Functions
}
